Move NSLayoutConstraint extensions
